:art: setup logger refactoring
