+++ light_agents/core/tool_registry.py
@@ -1,6 +1,6 @@
 from typing import Any, Callable, Dict, List
 
-from light_agents.logger_config import setup_logger
+from light_agents.core.logger_config import setup_logger
 from light_agents.schemas.tool_schema import ToolBaseSchema, ToolResponseSchema
 
 logger = setup_logger(__name__)
