+++ light_agents/serializers/tools/claude_tools_serializer.py
@@ -1,9 +1,11 @@
 from enum import Enum
 from typing import Any, Dict
 
-from light_agents.logger_config import setup_logger
+from light_agents.core.logger_config import setup_logger
 from light_agents.schemas import ToolBaseSchema, ToolUseMessage
-from light_agents.serializers.tools.base_serializers import python_type_to_json_type
+from light_agents.serializers.tools.base_serializers import (
+    python_type_to_json_type,
+)
 
 logger = setup_logger(__name__)
